backend/requirements.txt
# FastAPI stack
fastapi==0.118.0
uvicorn==0.32.0
pydantic==2.12.3
pydantic_core==2.41.4
requests==2.32.5
httpx==0.28.1
python-dotenv==1.0.1
bcrypt==4.3.0
argon2-cffi==25.1.0
aiosqlite==0.21.0
anyio==4.11.0
aiofiles==24.1.0
aiohttp==3.13.1
aiodns==3.5.0
arrow==1.3.0
certifi==2025.1.31
charset-normalizer==3.4.3
idna==3.10
urllib3==2.5.0
typing_extensions==4.15.0
click==8.1.8
itsdangerous==2.2.0
Jinja2==3.1.6
python-multipart==0.0.20
email_validator==2.2.0

# Django stack
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
numpy
pandas
paramiko
dj-database-url
python-dotenv
